create TopBar component
