+++ src/components/Routes.js
@@ -1,14 +1,14 @@
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import Dashboard from './Dashboard';
 import Home from './Home';
-import Tracks from './Tracks';
+import TracksPage from './TracksPage';
 
 const Routes = () => (
   <BrowserRouter>
     <Switch>
       <Route path="/" component={Home} exact />
       <Route path="/dashboard" component={Dashboard} />
-      <Route path="/tracks" component={Tracks} />
+      <Route path="/tracks" component={TracksPage} />
     </Switch>
   </BrowserRouter>
 );
