+++ src/components/TopBar.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const TopBar = (title) => (
+  <div className="top">
+    <p>{title}</p>
+  </div>
+);
+
+export default TopBar;
