+++ src/components/Tracks.js
@@ -1,36 +1,40 @@
-import React, { useEffect } from 'react';
-import axios from 'axios';
-import { useDispatch } from 'react-redux';
-import Logout from './Logout';
-import Navigation from './Navigation';
-import { userTracks } from '../redux/actions/actions';
-import Track from './Track';
+import React from 'react';
+import { useSelector } from 'react-redux';
 
 const Tracks = () => {
-  const dispatch = useDispatch();
-
-  const fetchData = async () => {
-    await axios.get('http://localhost:3001/tracks',
-      { withCredentials: true }).then((response) => {
-      if (response.data.registers) {
-        const tracks = response.data.registers;
-        dispatch(userTracks(tracks));
-      }
-    }).catch((err) => {
-      console.log('resgistration error', err);
+  const trackStored = useSelector((state) => state.tracks);
+  if (trackStored.payload && trackStored.payload.length > 0) {
+    return trackStored.payload.map((track) => {
+      const {
+        day, electrodomestic, id, watts,
+      } = track;
+      const timeConnected = track.time_connected;
+      return (
+        <div key={id}>
+          <p>
+            Date:
+            {day}
+          </p>
+          <p>
+            Electrodomestic:
+            {electrodomestic}
+          </p>
+          <p>
+            Time Connected:
+            {timeConnected}
+          </p>
+          <p>
+            Watts:
+            {watts}
+          </p>
+        </div>
+      );
     });
-  };
-
-  useEffect(() => {
-    fetchData();
-  }, []);
+  }
 
   return (
     <div>
-      <h1>Tracks</h1>
-      <Track />
-      <Navigation />
-      <Logout />
+      <p> </p>
     </div>
   );
 };
